Successful download test tarball, extract, and find tool.

+++ shedclient/model/tasks.py
@@ -4,6 +4,7 @@ import os
 import tempfile
 
 from celery import Celery
+from kombu import Queue
 
 from galaxy.tools.deps import commands
 
@@ -14,8 +15,20 @@ from shedclient import context
 
 APP_PATH = "%s.app" % __name__
 APP_NAME = "shedclient"
+
+
+CELERY_DEFAULT_QUEUE = 'default'
+CELERY_QUEUES = (
+    Queue('default', routing_key='task.#'),
+    Queue('toolpanel', routing_key='task.#'),
+)
+CELERY_DEFAULT_EXCHANGE = 'tasks'
+CELERY_DEFAULT_EXCHANGE_TYPE = 'topic'
+CELERY_DEFAULT_ROUTING_KEY = 'task.default'
+
 CELERY_ROUTES = {
-    'shedclient.modify_toolpanel': {'queue': 'toolpanel'}
+    'shedclient.install_from_url': {'queue': 'default'},
+    'shedclient.modify_toolpanel': {'queue': 'toolpanel'},
 }
 
 app = Celery(APP_NAME)
@@ -24,6 +37,11 @@ app.conf.update(
     CELERY_ACCEPT_CONTENT=['json'],
     CELERY_RESULT_SERIALIZER='json',
     CELERY_ROUTES=CELERY_ROUTES,
+    CELERY_DEFAULT_QUEUE=CELERY_DEFAULT_QUEUE,
+    CELERY_QUEUES=CELERY_QUEUES,
+    CELERY_DEFAULT_EXCHANGE=CELERY_DEFAULT_EXCHANGE,
+    CELERY_DEFAULT_EXCHANGE_TYPE=CELERY_DEFAULT_EXCHANGE_TYPE,
+    CELERY_DEFAULT_ROUTING_KEY=CELERY_DEFAULT_ROUTING_KEY,
 )
 
 
@@ -32,7 +50,16 @@ def install_from_url(shed_client_context, install_request):
     shed_client_context = context.ensure(shed_client_context)
     url = install_request["url"]
     installable_type = install_request["installable_type"]
-    target_directory = install_request["target_directory"]
+    installable_id = install_request["id"]
+    installable_version = install_request["version"]
+    installable_directory = shed_client_context.install_directory.installable_directory(
+        installable_type,
+        installable_id,
+        installable_version
+    )
+    target_directory = installable_directory.source_path
+    if os.path.exists(installable_directory.source_path):
+        raise Exception("Installable already exists.")
     download_and_extract_archive(url, target_directory)
     verify_download(installable_type, target_directory)
 
@@ -72,7 +99,9 @@ def download_installable_archive(url):
         if commands.shell(full_command):
             raise Exception("Failed to download shed tarball [%s]." % url)
     except Exception:
-        os.remove(temp_file)
+        if os.path.exists(temp_file):
+            os.remove(temp_file)
+        raise
     return temp_file
